出席率を表示するための処理を追加しました。

+++ src/dto/AttendanceInfo.java
@@ -8,10 +8,10 @@ public class AttendanceInfo {
 	private String info;
 	private int absence;
 	private int late;
-	private int attendanceRate;
+	private double attendanceRate;
 	private int publicFlag;
 
-	public AttendanceInfo(int sNumber, String name, String date, int time, String info, int absence, int late, int attendanceRate,
+	public AttendanceInfo(int sNumber, String name, String date, int time, String info, int absence, int late, double attendanceRate,
 			int publicFlag) {
 		super();
 		this.sNumber = sNumber;
@@ -24,61 +24,77 @@ public class AttendanceInfo {
 		this.attendanceRate = attendanceRate;
 		this.publicFlag = publicFlag;
 	}
+
 	public int getSNumber() {
 		return sNumber;
 	}
+
 	public void setSNumber(int sNumber) {
 		this.sNumber = sNumber;
 	}
+
 	public String getName() {
 		return name;
 	}
+
 	public void setName(String name) {
 		this.name = name;
 	}
+
 	public String getDate() {
 		return date;
 	}
+
 	public void setDate(String date) {
 		this.date = date;
 	}
+
 	public int getTime() {
 		return time;
 	}
+
 	public void setTime(int time) {
 		this.time = time;
 	}
+
 	public String getInfo() {
 		return info;
 	}
+
 	public void setInfo(String info) {
 		this.info = info;
 	}
+
 	public int getAbsence() {
 		return absence;
 	}
+
 	public void setAbsence(int absence) {
 		this.absence = absence;
 	}
+
 	public int getLate() {
 		return late;
 	}
+
 	public void setLate(int late) {
 		this.late = late;
 	}
-	public int getAttendanceRate() {
+
+	public double getAttendanceRate() {
 		return attendanceRate;
 	}
-	public void setAttendanceRate(int attendanceRate) {
+
+	public void setAttendanceRate(double attendanceRate) {
 		this.attendanceRate = attendanceRate;
 	}
+
 	public int getPublicFlag() {
 		return publicFlag;
 	}
+
 	public void setPublicFlag(int publicFlag) {
 		this.publicFlag = publicFlag;
 	}
-
-
 
 }
